ch(remove-route-folder): remove route folder
