+++ src/index.js
@@ -2,7 +2,6 @@
 import express from 'express';
 import cors from 'cors';
 import graphqlHTTP from 'express-graphql';
-import routes from './routes/index';
 import authentication from './middlewares/authentication';
 import schema from './graphql/schema';
 import graphQlResolvers from './graphql/resolvers';
@@ -19,7 +18,6 @@ app.use('/aylon', graphqlHTTP({
   rootValue: graphQlResolvers,
   graphiql: true,
 }));
-app.use('/', routes);
 
 const port = process.env.PORT || 3000;
 const server = app.listen(port, () => console.log(`listening to port ${port}....`));
