+++ src/validations/productValidation.js
@@ -1,20 +1,21 @@
+/* eslint-disable no-restricted-globals */
 import Joi from '@hapi/joi';
-import voca from 'voca';
+import Helper from '../helpers/helper';
 
-class ProductValidation {
-  static validateProductService(req) {
+const { validate } = Helper;
+
+class Validations {
+  static validateProduct(args) {
     const schema = {
       name: Joi.string().required().min(3).max(255),
       category: Joi.string().required().min(3).max(255),
       quantity: Joi.number().integer().required().min(1),
       price: Joi.number().integer().required().min(1),
+      product_id: Joi.number().integer().min(1),
     };
 
-    req.body.name = voca.camelCase(req.body.name).trim().toLowerCase();
-    req.body.category = voca.camelCase(req.body.category).trim().toLowerCase();
-
-    return Joi.validate(req.body, schema);
+    return validate(args, schema);
   }
 }
 
-export default ProductValidation;
+export default Validations;
